Fix name and add nickname to Auth Hash Schema
Auth Hash Schema (1.0 and Later) states name should be:
"The best display name known to the strategy. Usually a concatenation
of first and last name, but may also be an arbitrary designator or
nickname for some strategies"

Azure sends "name" which is displayName in (on-premise) AD. Use that
as name and "unique_name" as nickname.